requirements.txt
pyodbc>=4.0.39
pyyaml>=6.0
